+++ src/GafferUIModule/PathListingWidgetBinding.cpp
@@ -62,13 +62,10 @@
 #include "QtCore/QModelIndex"
 #include "QtCore/QVariant"
 
-#if QT_VERSION >= 0x050000
-	#include "QtWidgets/QTreeView"
-	#include "QtWidgets/QFileIconProvider"
-#else
-	#include "QtGui/QTreeView"
-	#include "QtGui/QFileIconProvider"
-#endif
+#include "QtWidgets/QTreeView"
+#include "QtWidgets/QFileIconProvider"
+
+#include <unordered_map>
 
 using namespace boost::python;
 using namespace boost::posix_time;
@@ -76,6 +73,54 @@ using namespace Gaffer;
 
 namespace
 {
+
+// QVariant does have `operator <`, but it's deprecated as it doesn't define a
+// total ordering, making it unsuitable for our purposes.
+// See https://doc.qt.io/qt-5/qvariant-obsolete.html#operator-lt.
+bool variantLess( const QVariant &left, const QVariant &right )
+{
+	// Lexicographical comparison, first on type and then on value.
+
+	if( left.userType() < right.userType() )
+	{
+		return true;
+	}
+	else if( right.userType() < left.userType() )
+	{
+		return false;
+	}
+
+	assert( left.userType() == right.userType() );
+
+	switch( left.userType() )
+	{
+		case QVariant::Invalid :
+			// Both values are invalid, making them equal.
+			return false;
+		case QVariant::Int :
+			return left.toInt() < right.toInt();
+		case QVariant::UInt :
+			return left.toUInt() < right.toUInt();
+		case QVariant::LongLong:
+			return left.toLongLong() < right.toLongLong();
+		case QVariant::ULongLong:
+			return left.toULongLong() < right.toULongLong();
+		case QMetaType::Float:
+			return left.toFloat() < right.toFloat();
+		case QVariant::Double:
+			return left.toDouble() < right.toDouble();
+		case QVariant::Char:
+			return left.toChar() < right.toChar();
+		case QVariant::Date:
+			return left.toDate() < right.toDate();
+		case QVariant::Time:
+			return left.toTime() < right.toTime();
+		case QVariant::DateTime:
+			return left.toDateTime() < right.toDateTime();
+		default :
+			return left.toString() < right.toString();
+	}
+}
 
 IECore::InternedString g_namePropertyName( "name" );
 
@@ -355,16 +400,11 @@ class PathModel : public QAbstractItemModel
 
 		PathModel( QObject *parent = nullptr )
 			:	QAbstractItemModel( parent ),
-				m_rootItem( new Item( nullptr, 0, nullptr ) ),
+				m_rootItem( new Item( nullptr, nullptr ) ),
 				m_flat( true ),
 				m_sortColumn( -1 ),
 				m_sortOrder( Qt::AscendingOrder )
 		{
-		}
-
-		~PathModel() override
-		{
-			delete m_rootItem;
 		}
 
 		///////////////////////////////////////////////////////////////////
@@ -373,8 +413,12 @@ class PathModel : public QAbstractItemModel
 
 		void setColumns( const std::vector<ColumnPtr> columns )
 		{
+			/// \todo Maintain persistent indices etc
+			/// using `m_rootItem->update()`.
+			beginResetModel();
 			m_columns = columns;
-			setRoot( getRoot() ); // force a rebuild of our items
+			m_rootItem = new Item( getRoot(), nullptr );
+			endResetModel();
 		}
 
 		const std::vector<ColumnPtr> &getColumns() const
@@ -389,10 +433,7 @@ class PathModel : public QAbstractItemModel
 
 		void setRoot( PathPtr root )
 		{
-			beginResetModel();
-			delete m_rootItem;
-			m_rootItem = new Item( root, 0, nullptr );
-			endResetModel();
+			m_rootItem->update( root, this );
 		}
 
 		void setFlat( bool flat )
@@ -454,17 +495,17 @@ class PathModel : public QAbstractItemModel
 			}
 
 			QModelIndex result;
-			Item *item = m_rootItem;
+			Item *item = m_rootItem.get();
 			for( size_t i = rootPath->names().size(); i < path.size(); ++i )
 			{
 				bool foundNextItem = false;
-				const std::vector<Item *> &childItems = item->childItems( this );
-				for( std::vector<Item *>::const_iterator it = childItems.begin(), eIt = childItems.end(); it != eIt; ++it )
+				const Item::ChildContainer &childItems = item->childItems( this );
+				for( auto it = childItems.begin(), eIt = childItems.end(); it != eIt; ++it )
 				{
 					if( (*it)->path()->names()[i] == path[i] )
 					{
 						result = index( it - childItems.begin(), 0, result );
-						item = *it;
+						item = it->get();
 						foundNextItem = true;
 						break;
 					}
@@ -492,7 +533,7 @@ class PathModel : public QAbstractItemModel
 				return result;
 			}
 
-			indicesForPathsWalk( m_rootItem, QModelIndex(), paths, result );
+			indicesForPathsWalk( m_rootItem.get(), QModelIndex(), paths, result );
 			return result;
 		}
 
@@ -508,7 +549,7 @@ class PathModel : public QAbstractItemModel
 			}
 
 			Item *item = static_cast<Item *>( index.internalPointer() );
-			return item->data( index.column(), role, m_columns );
+			return item->data( index.column(), role, this );
 		}
 
 		QVariant headerData( int section, Qt::Orientation orientation, int role = Qt::DisplayRole ) const override
@@ -522,11 +563,11 @@ class PathModel : public QAbstractItemModel
 
 		QModelIndex index( int row, int column, const QModelIndex &parentIndex = QModelIndex() ) const override
 		{
-			Item *parentItem = parentIndex.isValid() ? static_cast<Item *>( parentIndex.internalPointer() ) : m_rootItem;
+			Item *parentItem = parentIndex.isValid() ? static_cast<Item *>( parentIndex.internalPointer() ) : m_rootItem.get();
 
 			if( row >=0 and row < (int)parentItem->childItems( this ).size() and column >=0 and column < (int)m_columns.size() )
 			{
-				return createIndex( row, column, parentItem->childItems( this )[row] );
+				return createIndex( row, column, parentItem->childItems( this )[row].get() );
 			}
 			else
 			{
@@ -564,9 +605,8 @@ class PathModel : public QAbstractItemModel
 			}
 			else
 			{
-				parentItem = m_rootItem;
+				parentItem = m_rootItem.get();
 			}
-
 			return parentItem->childItems( this ).size();
 		}
 
@@ -582,37 +622,42 @@ class PathModel : public QAbstractItemModel
 		// sorting".
 		void sort( int column, Qt::SortOrder order = Qt::AscendingOrder ) override
 		{
-			m_sortColumn = column;
-			m_sortOrder = order;
-
-			if( m_sortColumn < 0 )
+			if( m_sortColumn == column && m_sortOrder == order )
 			{
 				return;
 			}
 
-			layoutAboutToBeChanged();
-			m_rootItem->sort( this );
-			layoutChanged();
+			m_sortColumn = column;
+			m_sortOrder = order;
+
+			m_rootItem->update( m_rootItem->path(), this );
 		}
 
 	private :
 
 		// A single item in the PathModel - stores a path and caches
 		// data extracted from it to provide the model content.
-		struct Item
+		struct Item : public IECore::RefCounted
 		{
 
-			Item( Gaffer::PathPtr path, int row, Item *parent )
-				:	m_path( path ), m_parent( parent ), m_row( row ), m_dataDone( false ), m_childItemsDone( false )
+			Item( Gaffer::PathPtr path, Item *parent )
+				:	m_path( path ),
+					m_parent( parent ),
+					m_row( -1 ), // Assigned true value in `updateChildItems()`
+					m_dataState( State::Unrequested ),
+					m_childItemsState( State::Unrequested )
 			{
 			}
 
-			~Item()
+			IE_CORE_DECLAREMEMBERPTR( Item )
+
+			void update( const Gaffer::PathPtr &path, PathModel *model )
 			{
-				for( std::vector<Item *>::const_iterator it = m_childItems.begin(), eIt = m_childItems.end(); it != eIt; ++it )
-				{
-					delete *it;
-				}
+				// This is just intended to be called on the root item by the
+				// PathModel when the path changes.
+				assert( !m_parent );
+				m_path = path;
+				updateWalk( model );
 			}
 
 			Gaffer::Path *path()
@@ -633,11 +678,22 @@ class PathModel : public QAbstractItemModel
 			// Returns the data for the specified column and role, using the provided
 			// Columns to generate it as necessary. The Item is responsible for caching
 			// the results of these queries internally.
-			QVariant data( int column, int role, const std::vector<ColumnPtr> &columns )
+			QVariant data( int column, int role, const PathModel *model )
 			{
-				// We generate data for all columns and roles at once, on the assumption
-				// that access to one is likely to indicate upcoming accesses to the others.
-				ensureData( columns );
+				// We generate data for all columns and roles at once, on the
+				// assumption that access to one is likely to indicate upcoming
+				// accesses to the others.
+				//
+				// Note : `data()` is called from public query methods of the
+				// PathModel. We don't call `updateData()` here if
+				// `m_childItemsState == Dirty` because we can't modify the
+				// model from within a query function.
+
+				if( m_dataState == State::Unrequested )
+				{
+					m_dataState = State::Requested;
+					updateData( const_cast<PathModel *>( model ) );
+				}
 
 				switch( role )
 				{
@@ -650,98 +706,63 @@ class PathModel : public QAbstractItemModel
 				}
 			}
 
-			std::vector<Item *> &childItems( const PathModel *model )
-			{
-				if( !m_childItemsDone && m_path )
-				{
-					std::vector<Gaffer::PathPtr> children;
-					try
-					{
-						m_path->children( children );
-					}
-					catch( const std::exception &e )
-					{
-						IECore::msg( IECore::Msg::Error, "PathListingWidget", e.what() );
-					}
+			using ChildContainer = std::vector<Ptr>;
 
-					for( std::vector<Gaffer::PathPtr>::const_iterator it = children.begin(), eIt = children.end(); it != eIt; ++it )
-					{
-						m_childItems.push_back( new Item( *it, it - children.begin(), this ) );
-					}
-					// If the model is sorted, then we need to apply that same
-					// sorting to the new items - see comment for PathModel::sort().
-					sort( model );
+			ChildContainer &childItems( const PathModel *model )
+			{
+				// Note : `childItems()` is called from public query methods of
+				// the PathModel. We don't call `updateChildItems()` here
+				// if `m_childItemsState == Dirty` because we can't modify the
+				// model from within a query function.
+				if( m_childItemsState == State::Unrequested )
+				{
+					m_childItemsState = State::Requested;
+					updateChildItems( const_cast<PathModel *>( model ) );
 				}
-				m_childItemsDone = true;
 				return m_childItems;
-			}
-
-			void sort( const PathModel *model )
-			{
-				if( model->m_sortColumn < 0 || model->m_sortColumn >= model->columnCount() )
-				{
-					return;
-				}
-
-				if( !m_childItems.size() )
-				{
-					return;
-				}
-
-				SortableItems sortableChildren;
-				sortableChildren.reserve( m_childItems.size() );
-				for( int i = 0, e = m_childItems.size(); i < e; ++i )
-				{
-					m_childItems[i]->ensureData( model->getColumns() );
-					sortableChildren.push_back( SortableItem( m_childItems[i], i ) );
-				}
-
-				std::sort( sortableChildren.begin(), sortableChildren.end(), Less( model->m_sortColumn ) );
-
-				const bool reverse = model->m_sortOrder == Qt::DescendingOrder;
-				QModelIndexList changedPersistentIndexesFrom, changedPersistentIndexesTo;
-				for( int i = 0, e = sortableChildren.size(); i < e; ++i )
-				{
-					int fromRow = sortableChildren[i].second;
-					int toRow = reverse ? e - i - 1 : i;
-					m_childItems[toRow] = sortableChildren[i].first;
-					for( int c = 0, ce = model->getColumns().size(); c < ce; ++c )
-					{
-						changedPersistentIndexesFrom.append( model->createIndex( fromRow, c, sortableChildren[i].first ) );
-						changedPersistentIndexesTo.append( model->createIndex( toRow, c, sortableChildren[i].first ) );
-					}
-				}
-
-				const_cast<PathModel *>( model )->changePersistentIndexList( changedPersistentIndexesFrom, changedPersistentIndexesTo );
-
-				for( std::vector<Item *>::const_iterator it = m_childItems.begin(), eIt = m_childItems.end(); it != eIt; ++it )
-				{
-					(*it)->sort( model );
-				}
 			}
 
 			private :
 
-				typedef std::pair<Item *, size_t> SortableItem;
-				typedef std::vector<SortableItem> SortableItems;
-
-				void ensureData( const std::vector<ColumnPtr> &columns )
+				void updateWalk( PathModel *model )
 				{
-					if( m_dataDone )
+					if( m_dataState == State::Clean )
+					{
+						m_dataState = State::Dirty;
+					}
+					if( m_childItemsState == State::Clean )
+					{
+						m_childItemsState = State::Dirty;
+					}
+					updateData( model );
+					updateChildItems( model );
+					for( const auto &child : m_childItems )
+					{
+						child->updateWalk( model );
+					}
+				}
+
+				void updateData( PathModel *model )
+				{
+					if( m_dataState == State::Clean || m_dataState == State::Unrequested )
 					{
 						return;
 					}
 
-					m_displayData.reserve( columns.size() );
-					m_decorationData.reserve( columns.size() );
-					for( int i = 0, e = columns.size(); i < e; ++i )
+					std::vector<QVariant> newDisplayData;
+					std::vector<QVariant> newDecorationData;
+
+					newDisplayData.reserve( model->m_columns.size() );
+					newDecorationData.reserve( model->m_columns.size() );
+
+					for( int i = 0, e = model->m_columns.size(); i < e; ++i )
 					{
 						QVariant displayData;
 						QVariant decorationData;
 						try
 						{
-							displayData = columns[i]->data( m_path.get(), Qt::DisplayRole );
-							decorationData = columns[i]->data( m_path.get(), Qt::DecorationRole );
+							displayData = model->m_columns[i]->data( m_path.get(), Qt::DisplayRole );
+							decorationData = model->m_columns[i]->data( m_path.get(), Qt::DecorationRole );
 						}
 						catch( const std::exception &e )
 						{
@@ -754,69 +775,201 @@ class PathModel : public QAbstractItemModel
 							IECore::msg( IECore::Msg::Warning, "PathListingWidget", "Unknown error" );
 						}
 
-						m_displayData.push_back( displayData );
-						m_decorationData.push_back( decorationData );
+						newDisplayData.push_back( displayData );
+						newDecorationData.push_back( decorationData );
 					}
 
-					m_dataDone = true;
-				}
-
-				struct Less
-				{
-					Less( int column )
-						:	m_column( column )
+					if( newDisplayData != m_displayData || newDecorationData != m_decorationData )
 					{
-					}
-
-					bool operator () ( const SortableItem &leftItem, const SortableItem &rightItem )
-					{
-						const QVariant &left = leftItem.first->m_displayData[m_column];
-						const QVariant &right = rightItem.first->m_displayData[m_column];
-						switch( left.userType() )
+						m_displayData.swap( newDisplayData );
+						m_decorationData.swap( newDecorationData );
+						if( m_dataState != State::Requested )
 						{
-							case QVariant::Invalid :
-								return right.type() != QVariant::Invalid;
-							case QVariant::Int :
-								return left.toInt() < right.toInt();
-							case QVariant::UInt :
-								return left.toUInt() < right.toUInt();
-							case QVariant::LongLong:
-								return left.toLongLong() < right.toLongLong();
-							case QVariant::ULongLong:
-								return left.toULongLong() < right.toULongLong();
-							case QMetaType::Float:
-								return left.toFloat() < right.toFloat();
-							case QVariant::Double:
-								return left.toDouble() < right.toDouble();
-							case QVariant::Char:
-								return left.toChar() < right.toChar();
-							case QVariant::Date:
-								return left.toDate() < right.toDate();
-							case QVariant::Time:
-								return left.toTime() < right.toTime();
-							case QVariant::DateTime:
-								return left.toDateTime() < right.toDateTime();
-							default :
-								return left.toString().compare( right.toString() ) < 0;
+							model->dataChanged( model->createIndex( m_row, 0, this ), model->createIndex( m_row, model->m_columns.size() - 1, this ) );
 						}
 					}
 
-					private :
+					m_dataState = State::Clean;
+				}
 
-						int m_column;
+				void updateChildItems( PathModel *model )
+				{
+					if( m_childItemsState == State::Unrequested || m_childItemsState == State::Clean )
+					{
+						return;
+					}
 
-				};
+					// Construct a new ChildContainer to replace our previous children.
+					// Where possible we reuse existing children instead of creating new
+					// ones.
+
+					ChildContainer newChildItems;
+					if( m_path )
+					{
+						std::vector<Gaffer::PathPtr> children;
+						try
+						{
+							m_path->children( children );
+						}
+						catch( const std::exception &e )
+						{
+							IECore::msg( IECore::Msg::Error, "PathListingWidget", e.what() );
+						}
+
+						std::unordered_map<IECore::InternedString, Item *> oldChildMap;
+						for( const auto &oldChild : m_childItems )
+						{
+							oldChildMap[oldChild->path()->names().back()] = oldChild.get();
+						}
+
+						for( auto it = children.begin(), eIt = children.end(); it != eIt; ++it )
+						{
+							auto oldIt = oldChildMap.find( (*it)->names().back() );
+							if( oldIt != oldChildMap.end() )
+							{
+								// Reuse previous item.
+								Ptr itemToReuse = oldIt->second;
+								itemToReuse->m_path = *it;
+								newChildItems.push_back( itemToReuse );
+							}
+							else
+							{
+								// Make new item.
+								newChildItems.push_back( new Item( *it, this ) );
+							}
+						}
+					}
+
+					// Sort the new container if necessary.
+
+					if( model->m_sortColumn >= 0 && model->m_sortColumn < (int)model->m_columns.size() )
+					{
+						for( const auto &childItem : newChildItems )
+						{
+							if( childItem->m_dataState == State::Unrequested )
+							{
+								childItem->m_dataState = State::Requested;
+							}
+							childItem->updateData( model );
+						}
+						std::sort(
+							newChildItems.begin(), newChildItems.end(),
+							[model] ( const Item::Ptr &left, const Item::Ptr &right ) {
+								const QVariant &l = left->m_displayData[model->m_sortColumn];
+								const QVariant &r = right->m_displayData[model->m_sortColumn];
+								return model->m_sortOrder == Qt::AscendingOrder ? variantLess( l, r ) : variantLess( r, l );
+							}
+						);
+					}
+
+					// Early out if nothing has changed.
+
+					if( newChildItems == m_childItems )
+					{
+						m_childItemsState = State::Clean;
+						return;
+					}
+
+					// If we had children before, figure out the mapping from old to new,
+					// so we can tell Qt about it. This is necessary so that persistent
+					// indices used to represent selection and expansion remain valid.
+
+					QModelIndexList changedPersistentIndexesFrom, changedPersistentIndexesTo;
+
+					std::unordered_map<IECore::InternedString, size_t> newChildMap;
+					for( size_t i = 0; i < newChildItems.size(); ++i )
+					{
+						newChildMap[newChildItems[i]->path()->names().back()] = i;
+					}
+
+					for( const auto &oldChild : m_childItems )
+					{
+						auto nIt = newChildMap.find( oldChild->m_path->names().back() );
+						if( nIt != newChildMap.end() )
+						{
+							const int toRow = nIt->second;
+							for( int c = 0, ce = model->getColumns().size(); c < ce; ++c )
+							{
+								changedPersistentIndexesFrom.append( model->createIndex( oldChild->row(), c, oldChild.get() ) );
+								changedPersistentIndexesTo.append( model->createIndex( toRow, c, newChildItems[toRow].get() ) );
+							}
+						}
+						else
+						{
+							oldChild->invalidateIndexes( model, changedPersistentIndexesFrom, changedPersistentIndexesTo );
+						}
+					}
+
+					// Apply the update. We have to mark ourselves clean _before_
+					// changing layout, to avoid recursion when Qt responds to
+					// `layoutAboutToBeChanged()`.
+
+					const bool emitLayoutChanged = m_childItemsState != State::Requested;
+					m_childItemsState = State::Clean;
+
+					if( emitLayoutChanged )
+					{
+						model->layoutAboutToBeChanged( { model->createIndex( row(), 0, this ) } );
+					}
+
+					m_childItems.swap( newChildItems );
+					for( size_t i = 0; i < m_childItems.size(); ++i )
+					{
+						m_childItems[i]->m_row = i;
+					}
+
+					if( emitLayoutChanged )
+					{
+						model->changePersistentIndexList( changedPersistentIndexesFrom, changedPersistentIndexesTo );
+						model->layoutChanged();
+					}
+				}
+
+				void invalidateIndexes( PathModel *model, QModelIndexList &from, QModelIndexList &to )
+				{
+					for( int c = 0, ce = model->getColumns().size(); c < ce; ++c )
+					{
+						from.append( model->createIndex( m_row, c, this ) );
+						to.append( QModelIndex() );
+					}
+					for( const auto &child : m_childItems )
+					{
+						child->invalidateIndexes( model, from, to );
+					}
+				}
 
 				Gaffer::PathPtr m_path;
 				Item *m_parent;
 				int m_row;
 
-				bool m_dataDone;
+				// State transitions :
+				//
+				// - Unrequested->Requested : When first queried.
+				// - Requested->Clean : When first updated.
+				// - Clean->Dirty : When path changes.
+				// - Dirty->Clean : On all subsequent updates.
+				enum class State
+				{
+					// Initial state. Not yet requested by clients
+					// of the model, therefore not yet computed, and not
+					// in need of consideration during recursive updates.
+					Unrequested,
+					// Has just been requested for the first time. Needs
+					// to be updated, but there is no need to emit change
+					// signals for the first update.
+					Requested,
+					// Computed and up to date.
+					Clean,
+					// Stale data that needs recomputing.
+					Dirty
+				};
+
+				State m_dataState;
 				std::vector<QVariant> m_displayData;
 				std::vector<QVariant> m_decorationData;
 
-				bool m_childItemsDone;
-				std::vector<Item *> m_childItems;
+				State m_childItemsState;
+				std::vector<Ptr> m_childItems;
 
 		};
 
@@ -841,11 +994,11 @@ class PathModel : public QAbstractItemModel
 			for( const auto &childItem : item->childItems( this ) )
 			{
 				const QModelIndex childIndex = index( row++, 0, itemIndex );
-				indicesForPathsWalk( childItem, childIndex, paths, indices );
+				indicesForPathsWalk( childItem.get(), childIndex, paths, indices );
 			}
 		}
 
-		Item *m_rootItem;
+		Item::Ptr m_rootItem;
 		bool m_flat;
 		std::vector<ColumnPtr> m_columns;
 		int m_sortColumn;